File structure setup script
This is a script designed for an Ubuntu server I made that adds the necessary file structure and conf entries for a new site you want to add.  This is specifically for a server that's running apache behind an nginx reverse proxy, but it can be tuned.

The permissions management may need some work as well.