(mt) Media Temple GRID site grabber
This is an expect script that takes the details of a GRID service and grabs all of the files in the html directory for a given site.  The only part that would need to be tuned for a different type of server may be the arguments it takes and what directory it connects to specifically.